+++ frontend/src/components/VoteScore/index.js
@@ -1,18 +1,38 @@
-import { connect } from 'react-redux';
-import { votePostById } from '../../redux/modules/posts';
-import { getPostVoteScore } from '../../redux/selectors/posts';
+import React from 'react';
+import T from 'prop-types';
+import { FaCaretSquareODown, FaCaretSquareOUp } from 'react-icons/lib/fa';
 
-import Container from './Container';
+import styles from './styles.css';
 
-const mapStateToProps = (state, ownProps) => {
-  const postId = ownProps.postId;
-  return {
-    postId,
-    voteScore: getPostVoteScore(state, postId),
-  };
+const VoteScore = (props) => {
+  const { score, onVoteUp, onVoteDown } = props;
+  return (
+    <div className={styles.voteScoreWrapper}>
+      <span
+        role="button"
+        className={styles.voteControlUp}
+        tabIndex="0"
+        onClick={onVoteUp}
+      >
+        <FaCaretSquareOUp />
+      </span>
+      <div className={`${styles.voteScore} ${score < 0 ? styles.negative : styles.positive}`}><span>{score}</span></div>
+      <span
+        role="button"
+        tabIndex="0"
+        className={styles.voteControlDown}
+        onClick={onVoteDown}
+      >
+        <FaCaretSquareODown />
+      </span>
+    </div>
+  );
 };
 
-export default connect(
-  mapStateToProps,
-  { votePostById }
-)(Container);
+VoteScore.propTypes = {
+  score: T.number.isRequired,
+  onVoteUp: T.func.isRequired,
+  onVoteDown: T.func.isRequired,
+};
+
+export default VoteScore;
